+++ src/components/homePage/catalog/checkout/FormWithNotifications.tsx
@@ -1,19 +1,34 @@
 "use client";
-import { useState } from "react";
-
+import { Dispatch, SetStateAction } from "react";
 import NotificationPopUp from "@/components/shared/notifications/NotificationPopUp";
-
 import CheckoutForm from "./CheckoutForm";
+import { FormikProps } from "formik";
+import { ValuesCheckoutFormType } from "./CheckoutPopUp";
 
-export default function FormWithNotifications() {
-  const [isError, setIsError] = useState(false);
-  const [isNotificationShown, setIsNotificationShown] = useState(false);
+interface FormWithNotificationsProps {
+  formik: FormikProps<ValuesCheckoutFormType>; // Тип для Formik пропсів
+  setIsLoading: Dispatch<SetStateAction<boolean>>;
+  setIsError: Dispatch<SetStateAction<boolean>>;
+  setIsNotificationShown: Dispatch<SetStateAction<boolean>>;
+  isError: boolean;
+  isNotificationShown: boolean;
+}
 
+export default function FormWithNotifications({
+  formik,
+  setIsLoading,
+  isError,
+  isNotificationShown,
+  setIsError,
+  setIsNotificationShown,
+}: FormWithNotificationsProps) {
   return (
     <>
       <CheckoutForm
+        formik={formik} // Передаємо formik як пропс в CheckoutForm
         setIsError={setIsError}
         setIsNotificationShown={setIsNotificationShown}
+        setIsLoading={setIsLoading}
       />
       <NotificationPopUp
         isNotificationShown={isNotificationShown}
