+++ src/components/shared/forms/formComponents/SubmitButton.tsx
@@ -3,6 +3,7 @@ import React from "react";
 import Button from "../../buttons/Button";
 
 interface SubmitButtonProps {
+  onClick: () => void | Promise<void>;
   dirty: boolean;
   isValid: boolean;
   isLoading: boolean;
@@ -11,6 +12,7 @@ interface SubmitButtonProps {
 }
 
 export default function SubmitButton({
+  onClick,
   dirty,
   isValid,
   isLoading,
@@ -20,6 +22,7 @@ export default function SubmitButton({
   return (
     <Button
       type="submit"
+      onClick={onClick}
       disabled={!(dirty && isValid) || isLoading}
       isLoading={isLoading}
       className={`mt-4 tab:mt-6 mr-auto ${className}`}
